refactor: moved add-completions subcommand into seperate file
